Revert ULONG to unsigned long long

+++ xbmc/filesystem/AudioBookFileDirectory.cpp
@@ -96,8 +96,8 @@ bool CAudioBookFileDirectory::GetDirectory(const CURL& url, CFileItemList& items
   }
 
   std::map<std::string, std::string> fileTags;
-  std::map<ULONG, std::map<std::string, std::string>> chapterTags;
-  std::vector<std::tuple<ULONG, std::string, double, double>> chapterOrder;
+  std::map<unsigned long long, std::map<std::string, std::string>> chapterTags;
+  std::vector<std::tuple<unsigned long long, std::string, double, double>> chapterOrder;
   if (!isAudioBook)
   {
     CMusicInfoTagLoaderMatroska::GetMatroskaMusicTags(url.Get(), fileTags, chapterTags,
